+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.0.0-beta.XX] - 2018-XX-XX
+## [2.0.0-beta.42] - 2018-02-17
 ### Added
 - Added Patreon link in Contribute admin page
 - Added Theme Code Injection functionality
@@ -27,5 +27,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.0.0-beta.11] - 2018-01-20
 - First beta release
 
-[2.0.0-beta.12]: https://github.com/Requarks/wiki/releases/tag/2.0.0-beta.12
+[2.0.0-beta.42]: https://github.com/Requarks/wiki/releases/tag/2.0.0-beta.42
 [2.0.0-beta.11]: https://github.com/Requarks/wiki/releases/tag/2.0.0-beta.11